backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
loguru
pytest
httpx
sqlalchemy
psycopg2-binary
pillow
numpy
python-multipart
tensorflow
